MyRestAPI/app/requirements.txt
Flask-RESTful>0.3.8
Flask>=2.2.2
gunicorn==20.0.4
python-dotenv==0.19.0
flask-apispec>=0.11.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-marshmallow==1.2.0
pymysql==1.1.0
pandas==2.2.1
jinja2